+++ assets/js/settingsGeneral.js
@@ -1272,6 +1272,8 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _wordpress_hooks__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! @wordpress/hooks */ "@wordpress/hooks");
 /* harmony import */ var _wordpress_hooks__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(_wordpress_hooks__WEBPACK_IMPORTED_MODULE_3__);
 /* harmony import */ var _DateOffsetPreview__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./DateOffsetPreview */ "./assets/jsx/components/DateOffsetPreview.jsx");
+/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! @wordpress/components */ "@wordpress/components");
+/* harmony import */ var _wordpress_components__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(_wordpress_components__WEBPACK_IMPORTED_MODULE_5__);
 function _slicedToArray(r, e) { return _arrayWithHoles(r) || _iterableToArrayLimit(r, e) || _unsupportedIterableToArray(r, e) || _nonIterableRest(); }
 function _nonIterableRest() { throw new TypeError("Invalid attempt to destructure non-iterable instance.\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method."); }
 function _unsupportedIterableToArray(r, a) { if (r) { if ("string" == typeof r) return _arrayLikeToArray(r, a); var t = {}.toString.call(r).slice(8, -1); return "Object" === t && r.constructor && (t = r.constructor.name), "Map" === t || "Set" === t ? Array.from(r) : "Arguments" === t || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(t) ? _arrayLikeToArray(r, a) : void 0; } }
@@ -1287,11 +1289,13 @@ function _arrayWithHoles(r) { if (Array.isArray(r)) return r; }
 
 
 
+
 var _wp = wp,
   apiFetch = _wp.apiFetch;
 var PanelRow = wp.components.PanelRow;
 var PostTypeSettingsPanel = function PostTypeSettingsPanel(props) {
   var originalExpireTypeList = props.expireTypeList[props.postType];
+  var isPro = props.isPro;
   var _useState = (0,_wordpress_element__WEBPACK_IMPORTED_MODULE_1__.useState)(props.settings.taxonomy),
     _useState2 = _slicedToArray(_useState, 2),
     postTypeTaxonomy = _useState2[0],
@@ -1353,6 +1357,20 @@ var PostTypeSettingsPanel = function PostTypeSettingsPanel(props) {
     hasPendingValidation = _useState30[0],
     setHasPendingValidation = _useState30[1];
   var offset = expireOffset ? expireOffset : props.settings.globalDefaultExpireOffset;
+  var HelpText = function HelpText(props) {
+    return /*#__PURE__*/React.createElement("p", {
+      className: "description"
+    }, props.children);
+  };
+  var FieldRow = function FieldRow(props) {
+    var className = 'publishpress-settings-field-row';
+    if (props.className) {
+      className += ' ' + props.className;
+    }
+    return /*#__PURE__*/React.createElement("div", {
+      className: className
+    }, props.children);
+  };
   var taxonomyRelatedActions = ['category', 'category-add', 'category-remove', 'category-remove-all'];
   var onChangeTaxonomy = function onChangeTaxonomy(value) {
     setPostTypeTaxonomy(value);
@@ -1547,6 +1565,35 @@ var PostTypeSettingsPanel = function PostTypeSettingsPanel(props) {
         onChange: onChangeEmailNotification
       })));
     }
+  }
+
+  // Add promotional fields for non-pro users
+  if (!isPro) {
+    // Custom statuses promotional field
+    settingsRows.push( /*#__PURE__*/React.createElement(___WEBPACK_IMPORTED_MODULE_0__.SettingRow, {
+      label: props.text.fieldCustomStatuses,
+      key: 'custom-statuses'
+    }, /*#__PURE__*/React.createElement(FieldRow, null, /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement("input", {
+      type: "checkbox",
+      disabled: true
+    }), props.text.fieldCustomStatusesLabel, /*#__PURE__*/React.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_5__.Tooltip, {
+      text: props.text.proFeatureTooltip
+    }, /*#__PURE__*/React.createElement("span", {
+      className: "dashicons dashicons-lock pp-pro-loc-icon"
+    }))))));
+
+    // Metadata scheduling promotional field
+    settingsRows.push( /*#__PURE__*/React.createElement(___WEBPACK_IMPORTED_MODULE_0__.SettingRow, {
+      label: props.text.fieldMetadataScheduling,
+      key: 'metadata_mapping'
+    }, /*#__PURE__*/React.createElement(FieldRow, null, /*#__PURE__*/React.createElement("div", null, /*#__PURE__*/React.createElement("input", {
+      type: "checkbox",
+      disabled: true
+    }), props.text.fieldMetadataSchedulingLabel, /*#__PURE__*/React.createElement(_wordpress_components__WEBPACK_IMPORTED_MODULE_5__.Tooltip, {
+      text: props.text.proFeatureTooltip
+    }, /*#__PURE__*/React.createElement("span", {
+      className: "dashicons dashicons-lock pp-pro-loc-icon"
+    }))), /*#__PURE__*/React.createElement(HelpText, null, props.text.fieldMetadataSchedulingDescription))));
   }
   settingsRows = (0,_wordpress_hooks__WEBPACK_IMPORTED_MODULE_3__.applyFilters)('expirationdate_settings_posttype', settingsRows, props, isActive, _wordpress_element__WEBPACK_IMPORTED_MODULE_1__.useState);
   var fieldSetClassNames = props.isVisible ? 'pe-settings-fieldset' : 'pe-settings-fieldset hidden';